Add week7 hw1

+++ src/week4/WETH.sol
@@ -11,9 +11,8 @@ contract WETH {
     string public symbol;
     uint8 public decimals;
     uint256 _totalSupply;
-    mapping (address => uint256) _balance;
-    mapping (address => mapping(address => uint256)) _allowance;
-
+    mapping(address => uint256) _balance;
+    mapping(address => mapping(address => uint256)) _allowance;
 
     constructor() {
         name = "Wrapped Ether";
@@ -37,7 +36,7 @@ contract WETH {
 
     function withdraw(uint256 amount) external {
         require(_balance[msg.sender] >= amount, "Insufficient balance");
-        
+
         _balance[msg.sender] -= amount;
         _totalSupply -= amount;
         payable(msg.sender).transfer(amount);
@@ -45,15 +44,15 @@ contract WETH {
         emit Transfer(address(this), msg.sender, amount);
     }
 
-    function totalSupply() public view returns (uint256){
+    function totalSupply() public view returns (uint256) {
         return _totalSupply;
     }
 
-    function balanceOf(address who) public view returns (uint256){
+    function balanceOf(address who) public view returns (uint256) {
         return _balance[who];
     }
 
-    function _transfer(address from, address to, uint value) internal{
+    function _transfer(address from, address to, uint256 value) internal {
         require(_balance[from] >= value, "Not enought token to transfer");
         require(to != address(0), "Send to 0x00 address");
         _balance[from] -= value;
@@ -61,26 +60,26 @@ contract WETH {
         emit Transfer(from, to, value);
     }
 
-    function transfer(address to, uint256 value) public returns (bool){
+    function transfer(address to, uint256 value) public returns (bool) {
         _transfer(msg.sender, to, value);
         return true;
     }
 
-    function _approve(address owner, address spender, uint256 value)internal {
+    function _approve(address owner, address spender, uint256 value) internal {
         _allowance[owner][spender] = value;
         emit Approval(owner, spender, value);
     }
 
-    function approve(address spender, uint256 value) public returns (bool){
+    function approve(address spender, uint256 value) public returns (bool) {
         _approve(msg.sender, spender, value);
         return true;
     }
 
-    function allowance(address owner, address spender) public view returns (uint256){
+    function allowance(address owner, address spender) public view returns (uint256) {
         return _allowance[owner][spender];
     }
 
-    function transferFrom(address from, address to, uint256 value) public returns (bool){
+    function transferFrom(address from, address to, uint256 value) public returns (bool) {
         require(_allowance[from][msg.sender] >= value, "Not enought allowance to transfer");
         _approve(msg.sender, to, value);
         _transfer(from, to, value);
